Adding mutex to object creations and stoping cars while train is on a railway

+++ RoadSimulator/Classes/Train/Train.cs
@@ -25,7 +25,7 @@ namespace RoadSimulator.Classes.Train
 
         public Point TrainPoint { get; set; }
 
-        public int TrainSpeed { get; set; }
+        public double TrainSpeed { get; set; }
 
 
         public Train()
